pcc/requirements.txt
# Core dependencies for Pied Piper 2.0 (Phase 1)
typer>=0.9.0
cbor2>=5.6.0
python-magic>=0.4.27
requests>=2.31.0
python-dotenv>=1.0.0
cryptography>=41.0.0

# IPFS upload via web3.storage
web3storage>=0.3.0

# Optional: For development and testing
pytest>=8.0.0


torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
tqdm>=4.65.0

pyzstd==0.16.2
rich==13.7.1
click==8.1.7
python-dotenv==1.0.1
cryptography==43.0.1
python-magic-bin==0.4.14
requests==2.32.3
pytest==8.3.2